cleanup comp

+++ DeltaEngine/Engine/UI/Brush/SnapToGridBrush.hpp
@@ -89,17 +89,17 @@ public:
 
 private:
 	void MouseToCanvase(Math::Point& mouse){
-		if (mouse.GetX() < _screenViewPort.position.GetX())
-			mouse.SetX(_screenViewPort.position.GetX());
+		if (static_cast<float>(mouse.GetX()) < _screenViewPort.position.GetX())
+			mouse.SetX(static_cast<int>(_screenViewPort.position.GetX()));
 		
-		if (mouse.GetY() < _screenViewPort.position.GetY())
-			mouse.SetY(_screenViewPort.position.GetY());
+		if (static_cast<float>(mouse.GetY()) < _screenViewPort.position.GetY())
+			mouse.SetY(static_cast<int>(_screenViewPort.position.GetY()));
 		
-		if (mouse.GetX() > _screenViewPort.scale.GetX())
-			mouse.SetX(_screenViewPort.scale.GetX());
+		if (static_cast<float>(mouse.GetX()) > _screenViewPort.scale.GetX())
+			mouse.SetX(static_cast<int>(_screenViewPort.scale.GetX()));
 		
-		if (mouse.GetY() > _screenViewPort.scale.GetY())
-			mouse.SetY(_screenViewPort.scale.GetY());
+		if (static_cast<float>(mouse.GetY()) > _screenViewPort.scale.GetY())
+			mouse.SetY(static_cast<int>(_screenViewPort.scale.GetY()));
 
 	}
